fix(test): fix frontend tests
